MBS-29 Work on shared library
* Refactor to share code between device and control panel
