+++ src/control_panel/core/__init__.py
@@ -1,0 +1,3 @@
+"""
+Control Panel Core Logic
+"""
