+++ src/control_panel/ui/__init__.py
@@ -1,0 +1,7 @@
+"""
+Control Panel UI Components
+"""
+
+from .timeline_editor import TimelineEditor
+
+__all__ = ['TimelineEditor']
