+++ src/device/core/__init__.py
@@ -1,0 +1,8 @@
+"""
+Device Core Logic
+"""
+
+from .timeline_processor import TimelineProcessor
+from .trials import *
+
+__all__ = ['TimelineProcessor']
