+++ src/shared/__init__.py
@@ -1,0 +1,11 @@
+"""
+Shared library for bbox_controller project
+Contains common constants, models, and utilities used by both control panel and device
+"""
+
+from .constants import *
+from .models import *
+from .communication import *
+from .test_management import *
+
+__version__ = "1.0.0"
